Added authentication for update and delete room and also for routings.

+++ studentsync/base/views.py
@@ -1,12 +1,16 @@
 from django.shortcuts import render,redirect
 from .models import Room,Topic
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
 from .forms import RoomForm
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect("home")
     if request.method=="POST":
         username=request.POST.get('username')
         password=request.POST.get('password')
@@ -44,6 +48,7 @@ def room(request,pk):
             
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm
     if request.method=="POST":
@@ -54,9 +59,14 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to modify!!')
+    
     if request.method=="POST":
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -65,9 +75,17 @@ def updateRoom(request,pk):
     context={"form":form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to delete!!')
     if request.method=="POST":
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',{'obj':room})
+
+@login_required(login_url='login')
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
